Refactor: - move some endpoints to global app.py from stats - move some tests to `test_controller` Updated: - make max size configurable
