+++ server/mergin/stats/config.py
@@ -6,9 +6,6 @@ import os
 from decouple import config
 
 
-config_dir = os.path.abspath(os.path.dirname(__file__))
-
-
 class Configuration(object):
     # send statistics about usage
     COLLECT_STATISTICS = config("COLLECT_STATISTICS", default=True, cast=bool)
@@ -20,9 +17,3 @@ class Configuration(object):
     STATISTICS_URL = config(
         "STATISTICS_URL", default="https://api.merginmaps.com/monitoring/v1"
     ).rstrip("/")
-    DIAGNOSTIC_LOGS_DIR = config(
-        "DIAGNOSTIC_LOGS_DIR",
-        default=os.path.join(
-            config_dir, os.pardir, os.pardir, os.pardir, "diagnostic_logs"
-        ),
-    )
